feat: add drafts pages, new api and components
