+++ packages/web/app/invoice/_lib/useGeneratePdf.ts
@@ -1,15 +1,13 @@
 import { usePostMutation } from '@/src/lib/usePostMutation';
 
-const baseUrl = process.env.NEXT_PUBLIC_API_HOST;
-if (!baseUrl) {
-  throw new Error('missing NEXT_PUBLIC_API_HOST');
-}
-const apiKey = process.env.NEXT_PUBLIC_API_KEY;
-if (!apiKey) {
-  throw new Error('missing NEXT_PUBLIC_API_KEY');
-}
-
 export function useGeneratePdfMutation(templateName: string) {
+  const baseUrl = process.env.NEXT_PUBLIC_API_HOST;
+  const apiKey = process.env.NEXT_PUBLIC_API_KEY;
+
+  if (!apiKey || !baseUrl) {
+      throw new Error('Miss the API key or API host');
+  }
+
   return usePostMutation<Record<string, any>, Blob>({
     endpoint: `/api/pdf/generate/${templateName}`,
     requestOptions: {
