+++ packages/web/app/invoice/helpers/currencyFormatter.ts
@@ -1,0 +1,6 @@
+export const currencyFormatter = (currency: string) =>
+    new Intl.NumberFormat(undefined, {
+      style: 'currency',
+      currency,
+      maximumFractionDigits: 2,
+    });
