+++ packages/web/app/invoice/helpers/money.ts
@@ -1,0 +1,4 @@
+export function money(n: number) {
+    if (!isFinite(n)) return 0;
+    return Math.round(n * 100) / 100;
+  }
